requirements.txt
absl-py==0.7.1
Appium-Python-Client==0.38
assertpy==0.14
astor==0.8.0
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
bugjar==0.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
configparser==3.5.0
coverage==4.5.3
crayons==0.1.2
cricket==0.2.5
cycler==0.10.0
duvet==0.1.2
Faker==1.0.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.7
imageai==2.0.3
ImageHash==4.0
imageio==2.5.0
imbox==0.9.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.2.5
malcontent==0.0.0
march==0.0.0
Markdown==3.1.1
matplotlib==3.1.1
more-itertools==4.3.0
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
playsound==1.2.2
pluggy==0.8.0
protobuf==3.9.1
py==1.7.0
Pygments==2.3.1
pyparsing==2.4.2
pypiwin32==223
pytest==4.0.1
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-repeat==0.8.0
python-dateutil==2.7.5
PyWavelets==1.0.3
pywin32==224
PyYAML==5.1.2
requests==2.20.1
scipy==1.3.0
selenium==3.141.0
Selenium-Screenshot==1.5.0
six==1.11.0
style==1.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
text-unidecode==1.2
tkreadonly==0.6.1
update==0.0.1
urllib3==1.24.1
Wave==0.0.2
webdriver-manager==1.7
Werkzeug==0.15.5
wrapt==1.11.2
xlrd==1.2.0
